requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
requests==2.31.0